+++ src/Foundation/resources/database/migrations/2026_05_05_094401_add_tax_category_id_to_shipping_methods.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class () extends Migration {
+    public function up()
+    {
+        Schema::table('shipping_methods', function (Blueprint $table) {
+            $table->bigInteger('tax_category_id')->unsigned()->nullable();
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('shipping_methods', function (Blueprint $table) {
+            $table->dropColumn('tax_category_id');
+        });
+    }
+};